requirements.txt
Django==1.11.5
django-modeladmin-reorder==0.2
django-wysiwyg-redactor==0.5.1
djangorestframework==3.6.4
pytz==2017.2
